Move sequence to text
