+++ torchnlp/encoders/text/__init__.py
@@ -1,31 +1,31 @@
-from torchnlp.encoders.sequence.character_encoder import CharacterEncoder
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_COPY_INDEX
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_COPY_TOKEN
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_EOS_INDEX
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_EOS_TOKEN
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_PADDING_INDEX
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_PADDING_TOKEN
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_RESERVED_TOKENS
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_SOS_INDEX
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_SOS_TOKEN
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_UNKNOWN_INDEX
-from torchnlp.encoders.sequence.default_reserved_tokens import DEFAULT_UNKNOWN_TOKEN
-from torchnlp.encoders.sequence.delimiter_encoder import DelimiterEncoder
-from torchnlp.encoders.sequence.moses_encoder import MosesEncoder
-from torchnlp.encoders.sequence.sequence_encoder import pad_batch
-from torchnlp.encoders.sequence.sequence_encoder import pad_tensor
-from torchnlp.encoders.sequence.sequence_encoder import SequenceEncoder
-from torchnlp.encoders.sequence.spacy_encoder import SpacyEncoder
-from torchnlp.encoders.sequence.static_tokenizer_encoder import StaticTokenizerEncoder
-from torchnlp.encoders.sequence.subword_encoder import SubwordEncoder
-from torchnlp.encoders.sequence.treebank_encoder import TreebankEncoder
-from torchnlp.encoders.sequence.whitespace_encoder import WhitespaceEncoder
+from torchnlp.encoders.text.character_encoder import CharacterEncoder
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_COPY_INDEX
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_COPY_TOKEN
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_EOS_INDEX
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_EOS_TOKEN
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_PADDING_INDEX
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_PADDING_TOKEN
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_RESERVED_TOKENS
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_SOS_INDEX
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_SOS_TOKEN
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_UNKNOWN_INDEX
+from torchnlp.encoders.text.default_reserved_tokens import DEFAULT_UNKNOWN_TOKEN
+from torchnlp.encoders.text.delimiter_encoder import DelimiterEncoder
+from torchnlp.encoders.text.moses_encoder import MosesEncoder
+from torchnlp.encoders.text.text_encoder import pad_batch
+from torchnlp.encoders.text.text_encoder import pad_tensor
+from torchnlp.encoders.text.text_encoder import TextEncoder
+from torchnlp.encoders.text.spacy_encoder import SpacyEncoder
+from torchnlp.encoders.text.static_tokenizer_encoder import StaticTokenizerEncoder
+from torchnlp.encoders.text.subword_encoder import SubwordEncoder
+from torchnlp.encoders.text.treebank_encoder import TreebankEncoder
+from torchnlp.encoders.text.whitespace_encoder import WhitespaceEncoder
 
 __all__ = [
     'CharacterEncoder', 'DEFAULT_COPY_INDEX', 'DEFAULT_COPY_TOKEN', 'DEFAULT_EOS_INDEX',
     'DEFAULT_EOS_TOKEN', 'DEFAULT_PADDING_INDEX', 'DEFAULT_PADDING_TOKEN',
     'DEFAULT_RESERVED_TOKENS', 'DEFAULT_SOS_INDEX', 'DEFAULT_SOS_TOKEN', 'DEFAULT_UNKNOWN_INDEX',
     'DEFAULT_UNKNOWN_TOKEN', 'DelimiterEncoder', 'MosesEncoder', 'pad_batch', 'pad_tensor',
-    'SequenceEncoder', 'SpacyEncoder', 'StaticTokenizerEncoder', 'SubwordEncoder',
-    'TreebankEncoder', 'WhitespaceEncoder'
+    'TextEncoder', 'SpacyEncoder', 'StaticTokenizerEncoder', 'SubwordEncoder', 'TreebankEncoder',
+    'WhitespaceEncoder'
 ]
